completed

+++ animal-chipization/src/main/java/com/example/animalchipization/service/implementation/AnimalVisitedLocationServiceImpl.java
@@ -5,7 +5,7 @@ import com.example.animalchipization.data.repository.AnimalVisitedLocationReposi
 import com.example.animalchipization.data.repository.LocationPointRepository;
 import com.example.animalchipization.exception.AnimalIsAlreadyAtThisPointException;
 import com.example.animalchipization.exception.AttemptAddingLocationToAnimalWithDeadStatusException;
-import com.example.animalchipization.exception.FirstLocationPointCoincidesWithChippingPointException;
+import com.example.animalchipization.exception.FirstLocationPointConcidesWithChippingPointException;
 import com.example.animalchipization.model.Animal;
 import com.example.animalchipization.model.AnimalVisitedLocation;
 import com.example.animalchipization.model.LocationPoint;
@@ -22,6 +22,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import static com.example.animalchipization.data.specification.AnimalVisitedLocationSpecification.*;
 
@@ -58,9 +59,9 @@ public class AnimalVisitedLocationServiceImpl implements AnimalVisitedLocationSe
         }
         List<AnimalVisitedLocation> visitedLocations = animal.getVisitedLocations();
         if (visitedLocations.size() == 0 && animal.getChippingLocation().equals(animalVisitedLocation.getLocationPoint())) {
-            throw new FirstLocationPointCoincidesWithChippingPointException();
+            throw new FirstLocationPointConcidesWithChippingPointException();
         }
-        if (visitedLocations.size() != 0
+        if (visitedLocations.size() > 0
                 && visitedLocations.get(visitedLocations.size() - 1).getLocationPoint()
                 .equals(animalVisitedLocation.getLocationPoint())) {
             throw new AnimalIsAlreadyAtThisPointException();
@@ -75,28 +76,43 @@ public class AnimalVisitedLocationServiceImpl implements AnimalVisitedLocationSe
         Animal animal = animalRepository.findById(animalId)
                 .orElseThrow(NoSuchElementException::new);
         AnimalVisitedLocation visitedLocationPoint = animalVisitedLocationRepository
-                .findByAnimal_IdAndLocationPoint_Id(animalId, visitedLocationPointId)
-                .orElseThrow(NoSuchElementException::new);
+                .findById(visitedLocationPointId).orElseThrow(NoSuchElementException::new);
         LocationPoint locationPoint = locationPointRepository
                 .findById(locationPointId).orElseThrow(NoSuchElementException::new);
 
-        ListIterator<LocationPoint> locationsIterator = animal.getVisitedLocations()
-                .stream().map(AnimalVisitedLocation::getLocationPoint).toList().listIterator();
-        if (locationPoint.equals(animal.getChippingLocation())
-                && locationPoint.equals(locationsIterator.next())) {
-            throw new FirstLocationPointCoincidesWithChippingPointException();
+        if (!animal.getVisitedLocations().contains(visitedLocationPoint)) {
+            throw new NoSuchElementException();
         }
-        if (visitedLocationPoint.getLocationPoint().equals(locationPoint)) {
-            throw new AnimalIsAlreadyAtThisPointException();
+
+        List<AnimalVisitedLocation> locations = animal.getVisitedLocations();
+        try {
+            if (locations.get(0).getLocationPoint().equals(visitedLocationPoint.getLocationPoint()) &&
+                    locationPoint.equals(animal.getChippingLocation())) {
+                throw new FirstLocationPointConcidesWithChippingPointException();
+            }
+        } catch (IndexOutOfBoundsException ignoredException) {
         }
-        locationsIterator.set(visitedLocationPoint.getLocationPoint());
-        if (locationsIterator.hasNext() && locationsIterator.next().equals(locationPoint)) {
-            throw new AnimalIsAlreadyAtThisPointException();
+        try {
+            if (visitedLocationPoint.getLocationPoint().equals(locationPoint)) {
+                throw new AnimalIsAlreadyAtThisPointException();
+            }
+        } catch (IndexOutOfBoundsException ignoredException) {
         }
-        locationsIterator.set(visitedLocationPoint.getLocationPoint());
-        if (locationsIterator.hasPrevious() && locationsIterator.previous().equals(locationPoint)) {
-            throw new AnimalIsAlreadyAtThisPointException();
+        try {
+            if (locations.get(locations.indexOf(visitedLocationPoint) - 1).getLocationPoint()
+                    .equals(locationPoint)) {
+                throw new AnimalIsAlreadyAtThisPointException();
+            }
+        } catch (IndexOutOfBoundsException ignoredException) {
         }
+        try {
+            if (locations.get(locations.indexOf(visitedLocationPoint) + 1).getLocationPoint()
+                    .equals(locationPoint)) {
+                throw new AnimalIsAlreadyAtThisPointException();
+            }
+        } catch (IndexOutOfBoundsException ignoredException) {
+        }
+
 
         visitedLocationPoint.setLocationPoint(locationPoint);
         return animalVisitedLocationRepository.save(visitedLocationPoint);
@@ -105,17 +121,17 @@ public class AnimalVisitedLocationServiceImpl implements AnimalVisitedLocationSe
     @Override
     public void deleteAnimalVisitedLocation(Long animalId, Long pointId) {
         Animal animal = animalRepository.findById(animalId).orElseThrow(NoSuchElementException::new);
-        AnimalVisitedLocation visitedLocationPoint = animalVisitedLocationRepository
-                .findByAnimal_IdAndLocationPoint_Id(animalId, pointId).orElseThrow(NoSuchElementException::new);
-
-        ListIterator<AnimalVisitedLocation> visitedLocationsIterator = animal.getVisitedLocations().listIterator();
-        visitedLocationsIterator.set(visitedLocationPoint);
-        if (visitedLocationsIterator.nextIndex() == 0 && visitedLocationsIterator.hasNext()
-                && visitedLocationsIterator.next().getLocationPoint().equals(animal.getChippingLocation())) {
-            animalVisitedLocationRepository.delete(visitedLocationsIterator.previous());
+        AnimalVisitedLocation visitedLocationPoint = animalVisitedLocationRepository.findById(pointId)
+                .orElseThrow(NoSuchElementException::new);
+        if (!visitedLocationPoint.getAnimal().getId().equals(animalId)) {
+            throw new NoSuchElementException();
+        }
+        List<AnimalVisitedLocation> visitedLocations = animal.getVisitedLocations();
+        if (visitedLocations.indexOf(visitedLocationPoint) == 0 && visitedLocations.size() >= 2
+                && visitedLocations.get(1).getLocationPoint().equals(animal.getChippingLocation())) {
+            animalVisitedLocationRepository.delete(visitedLocations.get(1));
         }
         animalVisitedLocationRepository.delete(visitedLocationPoint);
-
     }
 
 }
